gradient of generalized spillage implemented
